Eclipse source clean up

+++ src/main/java/com/firemerald/additionalplacements/client/gui/AbstractBetterSelectionList.java
@@ -219,8 +219,8 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
             resourcelocation,
             this.getX(),
             this.getY(),
-            (float)this.getRight(),
-            (float)(this.getBottom() + (int)this.getScrollAmount()),
+            this.getRight(),
+            this.getBottom() + (int)this.getScrollAmount(),
             this.getWidth(),
             this.getHeight(),
             32,
@@ -288,7 +288,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 	/**
 	 * Called when a mouse button is clicked within the GUI element.
 	 * <p>
-	 * 
+	 *
 	 * @return {@code true} if the event is consumed, {@code false} otherwise.
 	 * @param mouseX the X coordinate of the mouse.
 	 * @param mouseY the Y coordinate of the mouse.
@@ -328,7 +328,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 	/**
 	 * Called when a mouse button is released within the GUI element.
 	 * <p>
-	 * 
+	 *
 	 * @return {@code true} if the event is consumed, {@code false} otherwise.
 	 * @param mouseX the X coordinate of the mouse.
 	 * @param mouseY the Y coordinate of the mouse.
@@ -346,7 +346,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 	/**
 	 * Called when the mouse is dragged within the GUI element.
 	 * <p>
-	 * 
+	 *
 	 * @return {@code true} if the event is consumed, {@code false} otherwise.
 	 * @param mouseX the X coordinate of the mouse.
 	 * @param mouseY the Y coordinate of the mouse.
@@ -385,7 +385,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 
 	/**
 	 * Sets the focus state of the GUI element.
-	 * 
+	 *
 	 * @param pFocused the focused GUI element.
 	 */
 	@Override
@@ -455,7 +455,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 	/**
 	 * Checks if the given mouse coordinates are over the GUI element.
 	 * <p>
-	 * 
+	 *
 	 * @return {@code true} if the mouse is over the GUI element, {@code false}
 	 *         otherwise.
 	 * @param mouseX the X coordinate of the mouse.
@@ -574,7 +574,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 
 		/**
 		 * Sets the focus state of the GUI element.
-		 * 
+		 *
 		 * @param focused {@code true} to apply focus, {@code false} to remove focus
 		 */
 		@Override
@@ -599,7 +599,7 @@ public abstract class AbstractBetterSelectionList<E extends AbstractBetterSelect
 		/**
 		 * Checks if the given mouse coordinates are over the GUI element.
 		 * <p>
-		 * 
+		 *
 		 * @return {@code true} if the mouse is over the GUI element, {@code false}
 		 *         otherwise.
 		 * @param mouseX the X coordinate of the mouse.
